requirements.txt
streamlit==1.31.0
audio-recorder-streamlit==0.0.10
python-dotenv==1.0.1
langchain==0.3.0
langchain-openai==0.2.0
gTTS==2.5.3
openai==1.51.0